+++ database/seeders/DatabaseSeeder.php
@@ -11,5 +11,6 @@ final class DatabaseSeeder extends Seeder
         $this->call(RbacSeeder::class);
         $this->call(SaasSeeder::class);
         $this->call(CmsSeeder::class);
+        $this->call(MemberSettingsSeeder::class);
     }
 }